pyproject.toml
[tool.uv]
package = false

[project]
name = "almaeng"
version = "6.0.0"
description = "AI 쇼핑 도우미 - 자연어로 검색하면 최저가를 찾아주는 서비스"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "xid", email = "xid@daemon.dev" }]
keywords = ["django", "htmx", "ai", "gemini", "shopping", "price-comparison", "11st", "naver", "coupang"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.12,<3.13"

dependencies = [
    # ============================================
    # 1. Core Framework
    # ============================================
    "django>=5.1,<6.0",
    "django-ninja-extra>=0.30.8",  # API Framework
    # ============================================
    # 2. Frontend (Hypermedia-Driven)
    # ============================================
    "django-unfold",            # Modern Admin
    "django-components",        # Server-side Components
    "django-htmx",              # HTMX Integration
    "django-browser-reload",    # Hot Reload (DX)
    # ============================================
    # 3. Database
    # ============================================
    "psycopg[binary]",          # PostgreSQL Driver
    "django-lifecycle",         # Model Lifecycle Hooks
    # ============================================
    # 4. Async & Tasks (Phase 2)
    # ============================================
    # "taskiq",                   # Async Task Queue
    # "taskiq-redis>=1.2.1",      # Redis broker
    # 가격 모니터링은 Phase 2로 연기
    # ============================================
    # 5. AI (Google Gemini)
    # ============================================
    "google-genai>=1.56.0",     # Gemini API
    "pydantic>=2.0,<3.0",       # Schema Validation
    "pydantic-ai>=1.22.0",      # Pydantic AI
    # ============================================
    # 6. Shopping APIs
    # ============================================
    "httpx>=0.27.0",            # Async HTTP Client
    # ============================================
    # 7. Infrastructure
    # ============================================
    "django-environ",           # Environment Config
    "django-compressor",        # Asset Compression
    "django-redis>=5.0.0",      # Redis Cache
    "whitenoise",               # Static Files
    "uvicorn[standard]",        # ASGI Server (Dev)
    "granian>=2.6.0",           # ASGI Server (Prod)
]

[project.optional-dependencies]
# Vision AI for sports science applications
vision = [
    "mediapipe",
    "ultralytics",  # YOLO
    "opencv-python",
]

[project.urls]
Homepage = "https://almaeng.daemonx.cc"
Documentation = "https://github.com/daemonXid/almaeng#readme"
Repository = "https://github.com/daemonXid/almaeng"

[dependency-groups]
dev = [
    # --- Testing (Essential) ---
    "pytest",                   # Test Framework
    "pytest-django",            # Django Integration
    "pytest-asyncio",           # Async Tests
    # --- Debugging (Essential) ---
    "django-debug-toolbar",     # Debug Toolbar
    "django-extensions",        # Shell Plus
    "ipython",                  # Enhanced Shell
    # --- Code Quality (Essential) ---
    "ruff",                     # Linter + Formatter
]


# ============================================
# Code Quality Tools
# ============================================
[tool.ruff]
line-length = 120
target-version = "py312"
src = ["backend"]
exclude = ["**/migrations/**"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "RUF",    # Ruff-specific
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "E402",   # module level import not at top (used for re-exports)
    "E731",   # do not assign a lambda expression (used for fallbacks)
    "B008",   # do not perform function calls in argument defaults
    "UP046",  # Use type parameters (PEP 695 optional style)
    "UP047",  # Use type parameters for generic functions (PEP 695 optional)
    "RUF001", # Ambiguous unicode characters (intentional emoji usage)
    "RUF012", # Mutable class attributes (Django Meta class pattern)
]

[tool.ruff.lint.isort]
known-first-party = ["domains", "config"]

[tool.mypy]
python_version = "3.12"
mypy_path = "backend"
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true
strict = false

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
pythonpath = ["backend"]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = "--reuse-db -v"
asyncio_mode = "auto"

